+++ RedPandaIDE/compiler/executablerunner.cpp
@@ -84,6 +84,7 @@ void ExecutableRunner::run()
     mProcess = std::make_shared<QProcess>();
     mProcess->setProgram(mFilename);
     mProcess->setArguments(splitProcessCommand(mArguments));
+    qDebug()<<splitProcessCommand(mArguments);
     mProcess->setWorkingDirectory(mWorkDir);
     QProcessEnvironment env = QProcessEnvironment::systemEnvironment();
     QString path = env.value("PATH");